requirements.txt
streamlit
langchain
langchain-core
langchain-ollama
langchain-chroma
ollama
pandas
chromadb
pypdf
python-docx